+++ CHANGELOG.md
@@ -4,5 +4,7 @@ CHANGELOG
 unreleased:
 -----------
 
+- Add '-f' or '--fullscreen" options. It causes the JSON to be displayed on the
+  alternate buffer, in fullscreen.
 - Use 'q' or 'escape' to quit.
 - Argument parser improvement.